Changed int variables to long long to deal with large prime numbers

+++ DSA/ComputeR.cpp
@@ -1,6 +1,6 @@
 #include "ComputeR.h"
 #include "../ModularExponentiation.h"
 
-int computeR(int g, int k, int p, int q) {
+long long int computeR(long long int g, long long int k, long long int p, long long int q) {
     return modExp(g, k, p) % q;
 }
