+++ DSA/ComputeR.h
@@ -1,1 +1,1 @@
-int computeR(int g, int k, int p, int q);
+long long int computeR(long long int g, long long int k, long long int p, long long int q);
