+++ ModularInverse.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include "ModularInverse.h"
 
-int modInverse(int k, int q) {
-    for(int i = 1; i < q; i++) {
+long long int modInverse(long long int k, long long int q) {
+    for(long long int i = 1; i < q; i++) {
         if((k * i) % q == 1) {
             return i;
         }
